requirements.txt
# AlphaGPT - Python Dependencies
# Python 3.10+ required (recommended: 3.11)
#
# Installation:
#   pip3 install -r requirements.txt
#   # Or use: pip install -r requirements.txt (if pip points to Python 3)
#
# For experimental scripts (times.py, lord/experiment.py):
#   pip3 install -r requirements-optional.txt
#   # Or use: pip install -r requirements-optional.txt (if pip points to Python 3)

# ===== Core Dependencies (Required for main pipeline) =====

# Deep Learning & Numerical Computing
torch>=2.0.0          
numpy>=1.24.0         

# Data Processing & Database
pandas>=2.0.0         
sqlalchemy>=2.0.0     
asyncpg>=0.28.0       

# Async HTTP
aiohttp>=3.9.0        

# Environment & Configuration
python-dotenv>=1.0.0  

# Logging & Progress
loguru>=0.7.0         
tqdm>=4.66.0          

# US Stock Data
yfinance>=0.2.30

# Dashboard & Visualization
streamlit>=1.28.0     
plotly>=5.17.0        

# PostgreSQL
psycopg2-binary>=2.9.10

